# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AuditResult": {
        "properties": {
          "confidence": {
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "evidence_refs": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "metrics": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Metrics",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "selection_digest": {
            "title": "Selection Digest",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/AuditStage"
            },
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "selection_digest",
          "metrics",
          "stages",
          "findings",
          "evidence_refs",
          "confidence",
          "limitations"
        ],
        "title": "AuditResult",
        "type": "object"
      },
      "AuditStage": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "status": {
            "const": "completed",
            "default": "completed",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label"
        ],
        "title": "AuditStage",
        "type": "object"
      },
      "Body_preview_import_v1_imports_preview_post": {
        "properties": {
          "package": {
            "contentMediaType": "application/octet-stream",
            "title": "Package",
            "type": "string"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "package"
        ],
        "title": "Body_preview_import_v1_imports_preview_post",
        "type": "object"
      },
      "ComparabilityRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "run_a": {
            "$ref": "#/components/schemas/RunSnapshot"
          },
          "run_b": {
            "$ref": "#/components/schemas/RunSnapshot"
          },
          "selection_digest": {
            "minLength": 8,
            "title": "Selection Digest",
            "type": "string"
          }
        },
        "required": [
          "selection_digest",
          "question",
          "run_a",
          "run_b"
        ],
        "title": "ComparabilityRequest",
        "type": "object"
      },
      "EvidenceLevel": {
        "enum": [
          "recorded",
          "declared",
          "unknown"
        ],
        "title": "EvidenceLevel",
        "type": "string"
      },
      "EvidenceRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "json_pointer": {
            "title": "Json Pointer",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "level": {
            "$ref": "#/components/schemas/EvidenceLevel"
          },
          "observed_value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Value"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Hash"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "source_path",
          "json_pointer",
          "level",
          "label",
          "observed_value"
        ],
        "title": "EvidenceRef",
        "type": "object"
      },
      "Finding": {
        "properties": {
          "confidence": {
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "evidence_ref_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ref Ids",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "candidate_pool_mismatch",
              "evaluation_split_mismatch",
              "preprocessing_mismatch",
              "metric_definition_mismatch",
              "missing_evidence",
              "matched_condition",
              "source_conflict"
            ],
            "title": "Kind",
            "type": "string"
          },
          "severity": {
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "const": "detected",
            "default": "detected",
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "severity",
          "title",
          "summary",
          "confidence",
          "evidence_ref_ids"
        ],
        "title": "Finding",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImportPreview": {
        "properties": {
          "audit_readiness": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Audit Readiness",
            "type": "object"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/PreviewFile"
            },
            "title": "Files",
            "type": "array"
          },
          "preview_id": {
            "title": "Preview Id",
            "type": "string"
          },
          "proposed_artifacts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Proposed Artifacts",
            "type": "array"
          },
          "proposed_experiment": {
            "additionalProperties": true,
            "title": "Proposed Experiment",
            "type": "object"
          },
          "proposed_runs": {
            "items": {
              "$ref": "#/components/schemas/RunSnapshot"
            },
            "title": "Proposed Runs",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "preview_id",
          "digest",
          "files",
          "proposed_experiment",
          "proposed_runs",
          "proposed_artifacts",
          "warnings",
          "audit_readiness"
        ],
        "title": "ImportPreview",
        "type": "object"
      },
      "PreviewFile": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "proposed_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Run Id"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "path",
          "sha256",
          "kind"
        ],
        "title": "PreviewFile",
        "type": "object"
      },
      "RunSnapshot": {
        "additionalProperties": false,
        "properties": {
          "declared_candidate_count": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declared Candidate Count"
          },
          "evaluation_split": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Split"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metric_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Definition"
          },
          "metric_name": {
            "title": "Metric Name",
            "type": "string"
          },
          "metric_value": {
            "minimum": 0.0,
            "title": "Metric Value",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "preprocessing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preprocessing"
          },
          "recorded_candidate_count": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded Candidate Count"
          },
          "recorded_evaluation_split": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded Evaluation Split"
          },
          "recorded_metric_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded Metric Definition"
          },
          "recorded_preprocessing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded Preprocessing"
          },
          "source_hashes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Source Hashes",
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "metric_name",
          "metric_value"
        ],
        "title": "RunSnapshot",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "WebMCP Research Audit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/v1/audits/comparability": {
      "post": {
        "operationId": "comparability_v1_audits_comparability_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Comparability",
        "tags": [
          "audits"
        ]
      }
    },
    "/v1/imports/preview": {
      "post": {
        "operationId": "preview_import_v1_imports_preview_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_import_v1_imports_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Import",
        "tags": [
          "imports"
        ]
      }
    }
  }
}